requirements.txt
django==5.0.3
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
requests==2.25.1
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.0
pandas==2.2.1